test: extract GlobalFilter test to a separate file
